+++ test/Armada.Test.Unit/Suites/Database/EdgeCaseTests.cs
@@ -402,7 +402,7 @@ namespace Armada.Test.Unit.Suites.Database
                     Task<bool>[] claimTasks = new Task<bool>[parallelCount];
                     for (int i = 0; i < parallelCount; i++)
                     {
-                        claimTasks[i] = db.Captains.TryClaimAsync(captain.Id, mission.Id, dock.Id);
+                        claimTasks[i] = db.Captains.TryClaimAsync(Armada.Core.Constants.DefaultTenantId, captain.Id, mission.Id, dock.Id);
                     }
 
                     bool[] results = await Task.WhenAll(claimTasks);